homework_4/backend/requirements.txt
Django>=3.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
gunicorn
whitenoise
pandas
ta